pyproject.toml
[project]
name = "alpacalyzer"
version = "0.1.0"
description = "Alpacalyzer Algo Trading"
requires-python = ">=3.13"
license = { file = "LICENSE" }
authors = [
    { name = "Kim Rejström", email = "mr.rejstrom@gmail.com" },
]
readme = { file = "README.md", content-type = "text/markdown" }
dependencies = [
    "alpaca-py>=0.38.0",
    "colorama>=0.4.6",
    "finviz",
    "langchain-core>=0.3.45",
    "langgraph>=0.3.11",
    "numpy>=2.2.3",
    "openai>=1.61.1",
    "pandas>=2.2.3",
    "pandas-stubs>=2.2.3.241126",
    "pydantic-extra-types>=2.10.2",
    "pytest-mock>=3.14.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "ruff>=0.8.6",
    "schedule>=1.2.2",
    "pandas-ta>=0.3.14b0",
    "tabulate>=0.9.0",
    "ty>=0.0.1a12",
    "types-requests>=2.32.0.20241016",
    "websockets>=14.2",
    "yfinance>=0.2.54",
]

[project.scripts]
alpacalyzer = "alpacalyzer.cli:main"

[project.urls]
homepage = "https://github.com/smarlhens/python-boilerplate#readme"
repository = "https://github.com/smarlhens/python-boilerplate"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
  "pytest>=8.3.3",
  "pytest-cov>=5.0.0",
  "bandit>=1.8.0",
  "docformatter>=1.7.5",
  "ruff>=0.8.6",
  "ty>=0.0.1a11",
  "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
# Reduced verbosity for faster test runs (use -v for more detail when needed)
addopts = "-ra --strict-markers --strict-config"
testpaths = "tests"
# Enable parallel execution with: pytest -n auto
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
filterwarnings = [
    "ignore::DeprecationWarning:websockets.legacy",
]

[tool.pytest_asyncio]
default_fixture_scope = "function"

[tool.ruff]
extend-exclude = [
    "__pycache__",
    "build",
    "dist",
]
target-version = "py313"
line-length = 200
src = ["src", "tests"]

[tool.ruff.lint]
extend-select = [
    "C4",
    "D200",
    "D201",
    "D204",
    "D205",
    "D206",
    "D210",
    "D211",
    "D213",
    "D300",
    "D402",
    "D403",
    "D404",
    "D419",
    "E",
    "F",
    "G010",
    "I",
    "INP001",
    "N805",
    "PERF101",
    "PERF102",
    "PERF401",
    "PERF402",
    "PGH004",
    "PGH005",
    "PIE794",
    "PIE796",
    "PIE807",
    "PIE810",
    "RET502",
    "RET503",
    "RET504",
    "RET505",
    "RUF015",
    "RUF100",
    "UP",
    "W",
]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["S101"]

[tool.ty]

[tool.ty.environment]
python = ".venv"
root = ["src"]
python-version = "3.13"

[tool.ty.rules]
"unresolved-import" = "ignore"
"invalid-argument-type" = "warn"
"possibly-unresolved-reference" = "error"
"unsupported-operator" = "error"

[tool.uv.sources]
finviz = { git = "https://github.com/RiverRo/finviz.git", rev = "main" }
